# openapi/guild.v1alpha1.yaml
openapi: 3.1.0
info:
  title: Guild Control Plane API
  version: 0.1.0
  summary: Orchestrator-agnostic institution layer for AI teams.
  license:
    name: MIT
    identifier: MIT
  description: |
    Guild standardizes ownership, bounded handoffs, durable artifacts, and governed learning
    above agent orchestrators. This OpenAPI document describes the bootstrap control-plane API.
servers:
  - url: http://localhost:8080
    description: Local bootstrap server
security: []
...
paths:
...
  /api/v1/status:
    get:
      operationId: getStatus
      summary: Get bootstrap control-plane status.
      tags: [Status]
      responses:
        "200":
          description: Control-plane status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/taskpacks:
    get:
      operationId: listTaskpacks
      summary: List taskpacks.
      tags: [Taskpacks]
      responses:
        "200":
          description: Taskpack list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskpackList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createTaskpack
      summary: Create a taskpack.
      tags: [Taskpacks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Taskpack"
      responses:
        "201":
          description: Created taskpack.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taskpack"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/dri-bindings:
    get:
      operationId: listDriBindings
      summary: List DRI bindings.
      tags: [DRI Bindings]
      responses:
        "200":
          description: DRI binding list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriBindingList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createDriBinding
      summary: Create a DRI binding.
      tags: [DRI Bindings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriBinding"
      responses:
        "201":
          description: Created DRI binding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriBinding"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/artifacts:
    get:
      operationId: listArtifacts
      summary: List artifacts.
      tags: [Artifacts]
      responses:
        "200":
          description: Artifact list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createArtifact
      summary: Create an artifact metadata record.
      tags: [Artifacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Artifact"
      responses:
        "201":
          description: Created artifact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/promotion-records:
    get:
      operationId: listPromotionRecords
      summary: List promotion records.
      tags: [Promotion Records]
      responses:
        "200":
          description: Promotion record list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionRecordList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createPromotionRecord
      summary: Create a promotion record.
      tags: [Promotion Records]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionRecord"
      responses:
        "201":
          description: Created promotion record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/governance-policies:
    get:
      operationId: listGovernancePolicies
      summary: List governance policies.
      tags: [Institution]
      responses:
        "200":
          description: Governance policy list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernancePolicyList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createGovernancePolicy
      summary: Create a governance policy.
      tags: [Institution]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernancePolicy"
      responses:
        "201":
          description: Created governance policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernancePolicy"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/approval-requests:
    get:
      operationId: listApprovalRequests
      summary: List human approval requests.
      tags: [Institution]
      responses:
        "200":
          description: Approval request list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequestList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createApprovalRequest
      summary: Create a human approval request.
      tags: [Institution]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequest"
      responses:
        "201":
          description: Created approval request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/promotion-gates:
    get:
      operationId: listPromotionGates
      summary: List promotion gates.
      tags: [Institution]
      responses:
        "200":
          description: Promotion gate list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionGateList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createPromotionGate
      summary: Create a promotion gate.
      tags: [Institution]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionGate"
      responses:
        "201":
          description: Created promotion gate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionGate"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/commons-entries:
    get:
      operationId: listCommonsEntries
      summary: List accepted commons entries.
      tags: [Institution]
      responses:
        "200":
          description: Commons entry list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonsEntryList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createCommonsEntry
      summary: Create a commons registry entry.
      tags: [Institution]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommonsEntry"
      responses:
        "201":
          description: Created commons entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonsEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/evaluation-jobs:
    get:
      operationId: listEvaluationJobs
      summary: List replay/evaluation jobs.
      tags: [Replay]
      responses:
        "200":
          description: Evaluation job list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationJobList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: enqueueEvaluationJob
      summary: Queue a replay/evaluation suite job.
      tags: [Replay]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaySuite"
      responses:
        "202":
          description: Queued evaluation job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationJob"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
...
  responses:
    BadRequest:
      description: Request failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Server failed to complete the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UUID:
      type: string
      format: uuid
...
    StatusResponse:
      type: object
      additionalProperties: false
      required: [name, version, mode, ui_origin, services]
      properties:
        name:
          type: string
          enum: [guild]
        version:
          type: string
        mode:
          type: string
        ui_origin:
          type: string
          format: uri
        services:
          type: array
          items:
            type: object
            additionalProperties: true
...
    Taskpack:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - taskpack_id
        - title
        - objective
        - created_at
        - requested_by
        - priority
        - task_type
        - context_budget
        - permissions
        - acceptance_criteria
      properties:
        schema_version:
          type: string
          const: v1alpha1
        taskpack_id:
          $ref: "#/components/schemas/UUID"
        institution_id:
          $ref: "#/components/schemas/UUID"
        parent_taskpack_id:
          $ref: "#/components/schemas/UUID"
        title:
          $ref: "#/components/schemas/NonEmptyString"
        objective:
          $ref: "#/components/schemas/NonEmptyString"
        problem_statement:
          type: string
        task_type:
          type: string
          enum: [analysis, implementation, review, research, triage, evaluation, operations, custom]
        priority:
          type: string
          enum: [low, medium, high, critical]
        requested_by:
          $ref: "#/components/schemas/ActorRef"
        role_hint:
          type: string
          enum: [dri, explorer, builder, skeptic, reviewer, specialist, approver]
        inputs:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactRef"
        references:
          type: array
          items:
            $ref: "#/components/schemas/URI"
        labels:
          $ref: "#/components/schemas/Labels"
        context_budget:
          type: object
          additionalProperties: false
          required: [max_input_tokens, max_output_tokens, context_strategy]
          properties:
            max_input_tokens:
              type: integer
              minimum: 256
            max_output_tokens:
              type: integer
              minimum: 128
            context_strategy:
              type: string
              enum: [artifact_refs_first, summary_plus_refs, full_bundle]
            max_artifacts_inline:
              type: integer
              minimum: 0
        constraints:
          type: object
          additionalProperties: false
          properties:
            must_not:
              type: array
              items:
                type: string
            must_include:
              type: array
              items:
                type: string
            deadline_at:
              $ref: "#/components/schemas/Timestamp"
            cost_ceiling_usd:
              type: number
              minimum: 0
            tool_allowlist:
              type: array
              items:
                type: string
        permissions:
          type: object
          additionalProperties: false
          required: [allow_network, allow_shell, allow_external_write, approval_mode]
          properties:
            allow_network:
              type: boolean
            allow_shell:
              type: boolean
            allow_external_write:
              type: boolean
            approval_mode:
              type: string
              enum: [none, ask, required]
            scopes:
              type: array
              items:
                type: string
        acceptance_criteria:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [criterion_id, description]
            properties:
              criterion_id:
                type: string
                pattern: ^[a-z0-9][a-z0-9_-]{1,31}$
              description:
                $ref: "#/components/schemas/NonEmptyString"
              required:
                type: boolean
                default: true
              verification_hint:
                type: string
        delegation_policy:
          type: object
          additionalProperties: false
          properties:
            allow_subtasks:
              type: boolean
              default: false
            max_depth:
              type: integer
              minimum: 0
              default: 0
            max_parallel_subtasks:
              type: integer
              minimum: 1
              default: 1
        created_at:
          $ref: "#/components/schemas/Timestamp"
        trace_id:
          type: string
    DriBinding:
      type: object
      additionalProperties: false
      required: [schema_version, dri_binding_id, taskpack_id, owner, created_at, status]
      properties:
        schema_version:
          type: string
          const: v1alpha1
        dri_binding_id:
          $ref: "#/components/schemas/UUID"
        taskpack_id:
          $ref: "#/components/schemas/UUID"
        owner:
          $ref: "#/components/schemas/ActorRef"
        reviewers:
          type: array
          items:
            $ref: "#/components/schemas/ActorRef"
        specialists:
          type: array
          items:
            $ref: "#/components/schemas/ActorRef"
        approvers:
          type: array
          items:
            $ref: "#/components/schemas/ActorRef"
        escalation_policy:
          type: object
          additionalProperties: false
          properties:
            mode:
              type: string
              enum: [manual, time_based, policy_based]
            escalate_after_seconds:
              type: integer
              minimum: 0
            targets:
              type: array
              items:
                $ref: "#/components/schemas/ActorRef"
        approval_policy:
          type: object
          additionalProperties: false
          properties:
            required_for:
              type: array
              items:
                type: string
                enum: [external_write, network, prod_access, cost_overrun, publication]
            min_approvals:
              type: integer
              minimum: 0
        visibility_policy:
          type: object
          additionalProperties: false
          properties:
            mode:
              type: string
              enum: [team, project, institution, restricted]
            readers:
              type: array
              items:
                $ref: "#/components/schemas/ActorRef"
        status:
          type: string
          enum: [assigned, accepted, in_progress, blocked, completed, canceled]
        created_at:
          $ref: "#/components/schemas/Timestamp"
...
    TaskpackList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Taskpack"
    DriBindingList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DriBinding"
    ArtifactList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
    PromotionRecordList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PromotionRecord"
...
    GovernancePolicyList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GovernancePolicy"
    ApprovalRequestList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequest"
    PromotionGateList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PromotionGate"
    CommonsEntryList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CommonsEntry"